+++ CHANGELOG.md
@@ -20,6 +20,8 @@
 
 - Added nanosecond support for `pattern` and `Points.t`
 
+- Added `Timere.resolve_exn`
+
 ## Timedesc 0.3.1
 
 - Fixed `Timedesc.Span.make` handling of `ns` when `ns = Int.min_int`